Move FileUpload type from FileList to FileUploadField
FileUploadField owns the file upload state, so it is the more natural home for the type definition.
